Move Consumer API to writers
