+++ consumers/writers/api/logging.go
